pyproject.toml
[build-system]
requires = ["setuptools>=45"]
build-backend = "setuptools.build_meta"

[project]
name = "letoctf_cta_bot"
license = { text = "GPL-3.0-only" }
dependencies = [
    "aiogram",
    "aiogram_dialog",
    "python-dotenv",
    "sqlalchemy"
]
version = "0.0.1"

[tool.setuptools]
packages = ["cta"]


# [tool.setuptools.package-data]
# vkms = ["templates/*.html"]
